+++ ft_putptr_fd.c
@@ -6,7 +6,7 @@
 /*   By: tischmid <timo42@proton.me>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/05 16:51:47 by tischmid          #+#    #+#             */
-/*   Updated: 2023/10/05 21:44:25 by tischmid         ###   ########.fr       */
+/*   Updated: 2023/10/05 23:10:09 by tischmid         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,5 +18,5 @@ int	ft_putptr_fd(void *ptr, int fd)
 		return (ft_putstr_fd(NIL_PTR_STR, fd));
 	return (ft_putchar_fd('0', fd) + \
 			ft_putchar_fd('x', fd) + \
-			ft_puthex_fd((unsigned long long)ptr, 0, fd));
+			ft_puthex_fd((unsigned long)ptr, 0, fd));
 }
